[feat] 장바구니 주문 구현

+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductInfo.java
@@ -8,8 +8,8 @@ public record ProductInfo(
         String productName,
         String description,
         String brandId,
-        int price,
-        int stock,
+        Long price,
+        Long stock,
         String brandName,
         int likeCount,
         Boolean isLike
